refactor: Pieces 클래스 제거 & Team 인터페이스 추상 클래스로 변경
