+++ src/main/java/janggi/domain/side/Han.java
@@ -1,47 +1,47 @@
 package janggi.domain.side;
 
-import janggi.domain.Pieces;
 import janggi.domain.Position;
-import janggi.domain.piece.Piece;
-import janggi.dto.BoardSpot;
+import janggi.domain.piece.*;
 
+import java.util.HashMap;
 import java.util.Map;
-import java.util.Optional;
 
-public class Han implements Team {
+public class Han extends Team {
 
-    private final Pieces pieces;
-
-    private Han(Pieces pieces) {
-        this.pieces = pieces;
+    public Han(Map<Position, Piece> pieces) {
+        super(pieces);
     }
 
     public static Han createInitialHan() {
-        return new Han(Pieces.createHan());
-    }
-
-    @Override
-    public Map<Position, BoardSpot> makeSnapShot() {
-        return pieces.makeSnapShot();
-    }
-
-    @Override
-    public boolean isPieceExists(Position position) {
-        return pieces.isPieceExists(position);
-    }
-
-    @Override
-    public Optional<Piece> findPiece(Position position) {
-        return pieces.findPiece(position);
+        return new Han(initializePieces());
     }
 
     @Override
     public Team move(Position start, Position end) {
-        return new Han(pieces.move(start, end));
+        Map<Position, Piece> pieces = getPieces();
+        Piece piece = pieces.get(start);
+        Map<Position, Piece> updatedPieces = new HashMap<>(pieces);
+        updatedPieces.remove(start);
+        updatedPieces.put(end, piece);
+        return new Han(updatedPieces);
     }
 
     @Override
     public Team remove(Position position) {
-        return new Han(pieces.remove(position));
+        Map<Position, Piece> updatedPieces = new HashMap<>(getPieces());
+        updatedPieces.remove(position);
+        return new Han(updatedPieces);
+    }
+
+    private static Map<Position, Piece> initializePieces() {
+        Map<Position, Piece> pieces = new HashMap<>();
+        createChas(pieces, 10, TeamType.HAN);
+        createMas(pieces, 10, TeamType.HAN);
+        createSangs(pieces, 10, TeamType.HAN);
+        createSas(pieces, 10, TeamType.HAN);
+        createGung(pieces, 9, TeamType.HAN);
+        createPos(pieces, 8, TeamType.HAN);
+        createJols(pieces, 7, TeamType.HAN);
+        return pieces;
     }
 }
